fix scale type in g33.1

+++ src/emc/rs274ngc/interp_convert.cc
@@ -4444,7 +4444,7 @@ int Interp::convert_straight(int move,   //!< either G_0 or G_1
            (settings->spindle_turning != CANON_COUNTERCLOCKWISE)),
           _("Spindle not turning in G33.1"));
     START_SPEED_FEED_SYNCH(block->k_number, 0);
-    float scale = 1;
+    double scale = 1;
     if(block->i_flag){
         scale = block->i_number;
         if(scale < 1){
